feat: The Big Migration (#1011)
* migrate: Migrate the manifest to pixlet.

This commit migrates the app manifest to use the one defined in pixlet.

* migrate: Remove dead code.

This commit removes dead code that will no longer be needed after the
migration.

* migrate: Update the docs and the Makefile.

This commit updates the Makefile and docs to utilize pixlet instead.

* migrate: Add migrate command.

This commit adds a migration command that is mostly temporary, but it
will migrate from the old style manifest to the new one.

* migrate: Migrate to the new manifest style.

This commit migrates away from the old manifest style to the new one.

* migrate: Update apps.go to not require go modules.

This change allows app manifests to load without using go packages.

* migrate: Update build steps and remove dead code.

This commit updates the github build steps, upgrades pixlet to the
latest version, and removes dead code.

* Fixup pixlet check command